Added a help command and moved most of the tools and messages into their own header file.
